requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-amazon
boto3
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark
python-dotenv
spotipy
pyspark
findspark